Rename skills submodule to matchers
