+++ opsdroid/matchers.py
