+++ tests/test_skills.py
@@ -3,7 +3,7 @@ import unittest
 import unittest.mock as mock
 
 from opsdroid.core import OpsDroid
-from opsdroid import skills
+from opsdroid import matchers
 
 
 class TestSkillDecorators(unittest.TestCase):
@@ -13,7 +13,7 @@ class TestSkillDecorators(unittest.TestCase):
         with OpsDroid() as opsdroid:
             regex = r"(.*)"
             mockedskill = mock.MagicMock()
-            decorator = skills.match_regex(regex)
+            decorator = matchers.match_regex(regex)
             decorator(mockedskill)
             self.assertEqual(len(opsdroid.skills), 1)
             self.assertEqual(opsdroid.skills[0]["regex"], regex)
@@ -23,13 +23,13 @@ class TestSkillDecorators(unittest.TestCase):
         with OpsDroid() as opsdroid:
             action = "myaction"
             mockedskill = mock.MagicMock()
-            decorator = skills.match_apiai_action(action)
+            decorator = matchers.match_apiai_action(action)
             decorator(mockedskill)
             self.assertEqual(len(opsdroid.skills), 1)
             self.assertEqual(opsdroid.skills[0]["apiai_action"], action)
             self.assertIsInstance(opsdroid.skills[0]["skill"], mock.MagicMock)
             intent = "myIntent"
-            decorator = skills.match_apiai_intent(intent)
+            decorator = matchers.match_apiai_intent(intent)
             decorator(mockedskill)
             self.assertEqual(len(opsdroid.skills), 2)
             self.assertEqual(opsdroid.skills[1]["apiai_intent"], intent)
@@ -39,7 +39,7 @@ class TestSkillDecorators(unittest.TestCase):
         with OpsDroid() as opsdroid:
             crontab = "* * * * *"
             mockedskill = mock.MagicMock()
-            decorator = skills.match_crontab(crontab)
+            decorator = matchers.match_crontab(crontab)
             decorator(mockedskill)
             self.assertEqual(len(opsdroid.skills), 1)
             self.assertEqual(opsdroid.skills[0]["crontab"], crontab)
